[quant][fx] Don't assume bias is a keyword argument
`prepare_fx` currently assumes that bias is always a positional argument to
convolutions, and only a key argument to other functions. This happens to work
today due to a quirk in how `__torch_function__` is handled for python
functions but shouldn't be considered stable.

Instead, we should support `bias` for both positional and keyword forms.

ghstack-source-id: 76a19cca530f0c5ec56104d425b92a53dcafd002
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61647