[PyTorch Edge] Skip writing version during backport
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65842

During backport, only parts of the model (like bytecode.pkl) needs to be re-written, while the rest of the model is the same. However, `version` will always be re-written when `PyTorchStreamWriter` is destrcuted.

Change version to optional and add an api to allow skipping writing version when closing the writer.
ghstack-source-id: 139519393

Differential Revision: [D31262904](https://our.internmc.facebook.com/intern/diff/D31262904/)